clean up repr implementations and add repr for OptionalEventTime
