+++ raphtory-api/src/python/mod.rs
@@ -6,3 +6,5 @@ pub mod prop;
 pub mod timeindex;
 
 pub use prop::*;
+
+pub mod repr;
